make fastboot easy to use

+++ kernel/power/power.h
@@ -83,7 +83,7 @@ extern unsigned long reserved_size;
 extern int in_suspend;
 extern dev_t swsusp_resume_device;
 extern sector_t swsusp_resume_block;
-extern int fastboot;
+extern unsigned long fastboot;
 
 extern asmlinkage int swsusp_arch_suspend(void);
 extern asmlinkage int swsusp_arch_resume(void);
